Extract WinPE image internationalization from workspace prep
